refactor(ai/test): split war declaration target scoring tests by scenario (Refs #2289)
Replace monolithic war_declaration_target_scoring_test.dart with three
focused files: peacemaker relation threshold, warmonger weak-neighbor bonus,
and backstabber allied-target preference. No assertion or production changes.
